+++ src/basilisp/lang/compiler/generator.py
@@ -45,6 +45,8 @@ from basilisp.lang import symbol as sym
 from basilisp.lang import vector as vec
 from basilisp.lang.compiler.constants import (
     DEFAULT_COMPILER_FILE_PATH,
+    INTERFACE_KW,
+    REST_KW,
     SYM_DYNAMIC_META_KEY,
     SYM_NO_WARN_ON_REDEF_META_KEY,
     SYM_REDEF_META_KEY,
@@ -132,10 +134,6 @@ _THROW_PREFIX = "lisp_throw"
 _TRY_PREFIX = "lisp_try"
 _NS_VAR = "_NS"
 
-# Keyword constants used in generating code
-_INTERFACE_KW = kw.keyword("interface")
-_REST_KW = kw.keyword("rest")
-
 
 @attr.frozen
 class SymbolTableEntry:
@@ -209,7 +207,7 @@ class RecurType(Enum):
 class RecurPoint:
     loop_id: str
     type: RecurType
-    binding_names: Optional[Collection[str]] = None
+    binding_names: Optional[Iterable[str]] = None
     is_variadic: Optional[bool] = None
     has_recur: bool = False
 
@@ -273,7 +271,7 @@ class GeneratorContext:
             return False
 
     @property
-    def recur_point(self):
+    def recur_point(self) -> RecurPoint:
         return self._recur_points[-1]
 
     @contextlib.contextmanager
@@ -281,7 +279,7 @@ class GeneratorContext:
         self,
         loop_id: str,
         type_: RecurType,
-        binding_names: Optional[Collection[str]] = None,
+        binding_names: Optional[Iterable[str]] = None,
         is_variadic: Optional[bool] = None,
     ):
         self._recur_points.append(
@@ -305,7 +303,7 @@ class GeneratorContext:
             self._st.pop()
 
     @property
-    def current_this(self):
+    def current_this(self) -> sym.Symbol:
         return self._this[-1]
 
     @contextlib.contextmanager
@@ -1417,7 +1415,7 @@ def __deftype_or_reify_bases_to_py_ast(
                         ast.Call(
                             func=_NEW_KW_FN_NAME,
                             args=[
-                                ast.Constant(hash(_INTERFACE_KW)),
+                                ast.Constant(hash(INTERFACE_KW)),
                                 ast.Constant("interface"),
                             ],
                             keywords=[],
@@ -1695,7 +1693,7 @@ def __fn_decorator(
                                     ast.Call(
                                         func=_NEW_KW_FN_NAME,
                                         args=[
-                                            ast.Constant(hash(_REST_KW)),
+                                            ast.Constant(hash(REST_KW)),
                                             ast.Constant("rest"),
                                         ],
                                         keywords=[],
@@ -1810,11 +1808,7 @@ def __single_arity_fn_to_py_ast(  # pylint: disable=too-many-locals
                                     meta_decorators,
                                     [
                                         __fn_decorator(
-                                            (
-                                                (len(fn_args),)
-                                                if not method.is_variadic
-                                                else ()
-                                            ),
+                                            (len(fn_args),),
                                             has_rest_arg=method.is_variadic,
                                         )
                                     ],
@@ -1847,6 +1841,7 @@ def __multi_arity_dispatch_fn(  # pylint: disable=too-many-arguments,too-many-lo
     arity_map: Mapping[int, str],
     return_tags: Iterable[Optional[Node]],
     default_name: Optional[str] = None,
+    rest_arity_fixed_arity: Optional[int] = None,
     max_fixed_arity: Optional[int] = None,
     meta_node: Optional[MetaNode] = None,
     is_async: bool = False,
@@ -2013,7 +2008,16 @@ def __multi_arity_dispatch_fn(  # pylint: disable=too-many-arguments,too-many-lo
                             meta_decorators,
                             [
                                 __fn_decorator(
-                                    arity_map.keys(),
+                                    list(
+                                        chain(
+                                            arity_map.keys(),
+                                            (
+                                                [rest_arity_fixed_arity]
+                                                if rest_arity_fixed_arity is not None
+                                                else []
+                                            ),
+                                        )
+                                    ),
                                     has_rest_arg=default_name is not None,
                                 )
                             ],
@@ -2046,6 +2050,7 @@ def __multi_arity_fn_to_py_ast(  # pylint: disable=too-many-locals
 
     arity_to_name = {}
     rest_arity_name: Optional[str] = None
+    rest_arity_fixed_arity: Optional[int] = None
     fn_defs = []
     all_arity_def_deps: List[ast.AST] = []
     for arity in arities:
@@ -2054,6 +2059,7 @@ def __multi_arity_fn_to_py_ast(  # pylint: disable=too-many-locals
         )
         if arity.is_variadic:
             rest_arity_name = arity_name
+            rest_arity_fixed_arity = arity.fixed_arity
         else:
             arity_to_name[arity.fixed_arity] = arity_name
 
@@ -2109,6 +2115,7 @@ def __multi_arity_fn_to_py_ast(  # pylint: disable=too-many-locals
         arity_to_name,
         return_tags=[arity.tag for arity in arities],
         default_name=rest_arity_name,
+        rest_arity_fixed_arity=rest_arity_fixed_arity,
         max_fixed_arity=node.max_fixed_arity,
         meta_node=meta_node,
         is_async=node.is_async,
@@ -2567,6 +2574,7 @@ def __loop_recur_to_py_ast(
 ) -> GeneratedPyAST[ast.expr]:
     """Return a Python AST node for `recur` occurring inside a `loop`."""
     assert node.op == NodeOp.RECUR
+    assert ctx.recur_point.binding_names is not None
 
     recur_deps: List[ast.AST] = []
     recur_targets: List[ast.Name] = []
@@ -3223,7 +3231,6 @@ def _interop_prop_to_py_ast(
     assert node.op == NodeOp.HOST_FIELD
 
     target_ast = gen_py_ast(ctx, node.target)
-    assert not target_ast.dependencies
 
     return GeneratedPyAST(
         node=ast.Attribute(