+++ src/basilisp/lang/runtime.py
@@ -61,7 +61,7 @@ from basilisp.lang.interfaces import (
     ITransientSet,
 )
 from basilisp.lang.reference import RefBase, ReferenceBase
-from basilisp.lang.typing import CompilerOpts, LispNumber
+from basilisp.lang.typing import BasilispFunction, CompilerOpts, LispNumber
 from basilisp.lang.util import OBJECT_DUNDER_METHODS, demunge, is_abstract, munge
 from basilisp.util import Maybe
 
@@ -1286,12 +1286,46 @@ def _conj_ipersistentcollection(coll: IPersistentCollection, *xs):
     return coll.cons(*xs)
 
 
+def _update_signature_for_partial(f: BasilispFunction, num_args: int) -> None:
+    """Update the various properties of a Basilisp function for wrapped partials.
+
+    Partial applications change the number of arities a function appears to have.
+    This function computes the new `arities` set for the partial function by removing
+    any now-invalid fixed arities from the original function's set.
+
+    Additionally, partials do not take the meta from the wrapped function, so that
+    value should be cleared and the `with-meta` method should be replaced with a
+    new method."""
+    existing_arities: IPersistentSet[Union[kw.Keyword, int]] = f.arities
+    new_arities: Set[Union[kw.Keyword, int]] = set()
+    for arity in existing_arities:
+        if isinstance(arity, kw.Keyword):
+            new_arities.add(arity)
+        elif arity > num_args:
+            new_arities.add(arity - num_args)
+    if not new_arities:
+        if num_args in existing_arities:
+            new_arities.add(0)
+        else:
+            logger.warning(
+                f"invalid partial function application of '{f.__name__}' detected: "  # type: ignore[attr-defined]
+                f"{num_args} arguments given; expected any of: "
+                f"{', '.join(sorted(map(str, existing_arities)))}"
+            )
+    f.arities = lset.set(new_arities)
+    f.meta = None
+    f.with_meta = partial(_fn_with_meta, f)  # type: ignore[method-assign]
+
+
 def partial(f, *args, **kwargs):
     """Return a function which is the partial application of f with args and kwargs."""
 
     @functools.wraps(f)
     def partial_f(*inner_args, **inner_kwargs):
-        return f(*itertools.chain(args, inner_args), **{**kwargs, **inner_kwargs})
+        return f(*args, *inner_args, **{**kwargs, **inner_kwargs})
+
+    if hasattr(partial_f, "_basilisp_fn"):
+        _update_signature_for_partial(cast(BasilispFunction, partial_f), len(args))
 
     return partial_f
 
@@ -1760,11 +1794,13 @@ def _fn_with_meta(f, meta: Optional[lmap.PersistentMap]):
     return wrapped_f
 
 
-def _basilisp_fn(arities: Tuple[Union[int, kw.Keyword]]):
+def _basilisp_fn(
+    arities: Tuple[Union[int, kw.Keyword], ...]
+) -> Callable[..., BasilispFunction]:
     """Create a Basilisp function, setting meta and supplying a with_meta
     method implementation."""
 
-    def wrap_fn(f):
+    def wrap_fn(f) -> BasilispFunction:
         assert not hasattr(f, "meta")
         f._basilisp_fn = True
         f.arities = lset.set(arities)